starting to refactor bootstrap.js
