+++ .meteor/versions
@@ -59,6 +59,7 @@ less@1.0.14
 livedata@1.0.13
 localstorage@1.0.3
 logging@1.0.7
+medbook:patient-id@0.0.1
 medbook:testing-accounts@1.0.4
 meteor@1.1.6
 meteor-platform@1.2.2
