+++ README.md
@@ -84,14 +84,28 @@ else
 - Why are we attaching collections to the window[] object???  :(
 - What happens when list of patient outgrows what can display on a drop-down menu?
 - Should publish/subscribes be kept in lib/Patient_ID?  (hint: no)
-- Join
+- Why are we attaching things to the window?
 
 ==================================
 #### Architecture Recommendations  
 - Replace Select2 with a modal dialog?
+- Consolidate PatientEnrollment and Demographics collections
+- Consolidate Biopsy_V3 and Biopsy_Research collections
 
 
 ==================================
 #### Bugs  
 - autoform:autoformDate doesn't seem to be supported?  Should either implement or remove from schema.
 - datepickers don't seem to be working in the forms
+
+
+==================================
+#### Data Schema
+
+Schema Defines:
+- table
+- form
+Schema Name Defines:
+- menu item
+- url
+- collection name
