+++ packages/patient-id/patient-id-tests.js
@@ -1,0 +1,5 @@
+// Write your tests here!
+// Here is an example.
+Tinytest.add('example', function (test) {
+  test.equal(true, true);
+});
