+++ tests/nightwatch/commands/actions/completePatientEnrollmentForm.js
@@ -9,10 +9,30 @@ exports.command = function(patientRecord) {
     if(patientRecord){
 
       this
+        .clearValue('input[name="Patient_ID"]')
+        .clearValue('select[name="Study_Site"]')
+        .clearValue('input[name="Baseline_Sample_ID"]')
+        .clearValue('select[name="Baseline_Biopsy_Site"]')
+        .clearValue('input[name="Progression_Sample_ID"]')
+        .clearValue('select[name="Progression_Biopsy_Site"]')
+        .pause(500)
+
         .setValue('input[name="Patient_ID"]', patientRecord.patientId)
         .setValue('select[name="Study_Site"]', patientRecord.institution)
 
+        .clearValue('input[name="Baseline_Biopsy_Date"]')
+        .setValue('input[name="Baseline_Sample_ID"]', patientRecord.baselineSampleId)
+
+        .setValue('input[name="Baseline_Biopsy_Date"]', patientRecord.baselineBiopsyDate).pause(200)
+        .setValue('select[name="Baseline_Biopsy_Site"]', patientRecord.baselineBiopsySite)
+
+        .setValue('input[name="Progression_Sample_ID"]', patientRecord.progressionSampleId)
+        .clearValue('input[name="Progression_Biopsy_Date"]')
+        .setValue('input[name="Progression_Biopsy_Date"]', patientRecord.progressionBiopsyDate).pause(200)
+        .setValue('select[name="Progression_Biopsy_Site"]', patientRecord.progressionBiopsySite)
+
         .verify.elementPresent('button[type="submit"]')
+        .click('button[type="submit"]').pause(1000)
 
     }
