+++ tests/nightwatch/commands/components/reviewPatientEnrollmentRow.js
@@ -1,21 +1,30 @@
-    /*var patientEnrollmentRecord = {
-      patientId: "DTB-999",
-      institution: ""
-    }*/
+    /*
+      var patientEnrollmentRecord = {
+        patientId: "DTB-999",
+        institution: "OHSU",
+        baselineSampleId: "123",
+        baselineBiopsyDate: "2015-01-01",
+        baselineBiopsySite: "Bone"
+      }
+    */
 
 
 exports.command = function(patientEnrollmentRecord, rowIndex) {
   this
-    .verify.elementPresent("table thead tr")
-    .verify.elementPresent("table thead tr:nth-child(" + rowIndex + ")")
+    .verify.elementPresent("table tbody tr")
+    .verify.elementPresent("table tbody tr:nth-child(" + rowIndex + ")")
 
     if(patientEnrollmentRecord){
       this
-        .verify.elementPresent("table thead tr:nth-child(" + rowIndex + ") .Patient_ID")
-        .verify.elementPresent("table thead tr:nth-child(" + rowIndex + ") .Study_Site")
+        .verify.elementPresent("table tbody tr:nth-child(" + rowIndex + ") .Patient_ID")
+        .verify.elementPresent("table tbody tr:nth-child(" + rowIndex + ") .Study_Site")
 
-        .verify.containsText("table thead tr:nth-child(" + rowIndex + ") .Patient_ID", patientEnrollmentRecord.patientId )
-        .verify.containsText("table thead tr:nth-child(" + rowIndex + ") .Study_Site", patientEnrollmentRecord.institution )
+        // ISSUE #9 - this fails because m-dash is getting replaced with n-dash
+        //.verify.value("table tbody tr:nth-child(" + rowIndex + ") .Patient_ID span", patientEnrollmentRecord.patientId )
+        .verify.containsText("table tbody tr:nth-child(" + rowIndex + ") .Study_Site", patientEnrollmentRecord.institution )
+        .verify.containsText("table tbody tr:nth-child(" + rowIndex + ") .Baseline_Sample_ID", patientEnrollmentRecord.baselineSampleId )
+        .verify.containsText("table tbody tr:nth-child(" + rowIndex + ") .Baseline_Biopsy_Date", patientEnrollmentRecord.baselineBiopsyDate )
+        .verify.containsText("table tbody tr:nth-child(" + rowIndex + ") .Baseline_Biopsy_Site", patientEnrollmentRecord.baselineBiopsySite )
     }
 
   return this;
