+++ tests/nightwatch/reports/CHROME_43.0.2357.130_MAC_appLayout.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<testsuites errors="0"
+            failures="0"
+            tests="1">
+
+  <testsuite name="appLayout"
+    errors="0" failures="0" hostname="" id="" package="appLayout" skipped="0"
+    tests="1" time="4.329" timestamp="Thu, 25 Jun 2015 03:16:51 GMT">
+  
+    <testcase name="Layout &amp; Static Pages" time="4.329" assertions="4">
+
+    
+
+    
+
+    
+
+    
+    </testcase>
+      <system-err>
+        Unable to locate element: &quot;table tbody tr:nth-child(2) .Patient_ID span&quot; using: css selector
+      </system-err>
+  
+  
+  </testsuite>
+</testsuites>
