+++ tests/nightwatch/reports/FIREFOX_35.0_MAC_enrollPatient.xml
@@ -1,0 +1,83 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<testsuites errors="0"
+            failures="0"
+            tests="1">
+
+  <testsuite name="enrollPatient"
+    errors="0" failures="0" hostname="" id="" package="enrollPatient" skipped="0"
+    tests="1" time="19.35" timestamp="Fri, 26 Jun 2015 19:35:01 GMT">
+  
+    <testcase name="Patient Enrollment Form" time="19.35" assertions="34">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    </testcase>
+  
+  
+  </testsuite>
+</testsuites>
