+++ tests/nightwatch/walkthroughs/formsReadonly.js
@@ -1,0 +1,122 @@
+
+
+module.exports = {
+  "Readonly Forms Walkthrough" : function (client) {
+    client
+      .url("http://localhost:3000/CRF")
+      .resizeWindow(1024, 1200)
+      .reviewMainLayout()
+      .reviewSignInPage()
+      .signIn("house@test.org","house")
+
+      .sectionBreak("Clinical Info")
+      .selectFromSidebar('#Clinical_InfoButton')
+      .reviewPageTitle("Clinical_Info")
+      .reviewClinicalInfoForm()
+
+      .sectionBreak("Biopsy")
+      .selectFromSidebar('#SU2C_Biopsy_V3Button')
+      .reviewPageTitle("SU2C_Biopsy_V3")
+      .reviewBiopsyForm()
+
+      .sectionBreak("Follow Up")
+      .selectFromSidebar('#FollowupButton')
+      .reviewPageTitle("Followup")
+      .reviewFollowUpForm()
+
+      .sectionBreak("Subsequent Treatment")
+      .selectFromSidebar('#SU2C_Subsequent_Treatment_V1Button')
+      .reviewPageTitle("SU2C_Subsequent_Treatment_V1")
+      .reviewSubsequentTreatmentForm()
+
+      .sectionBreak("Prior Treatment")
+      .selectFromSidebar('#SU2C_Prior_TX_V3Button')
+      .reviewPageTitle("SU2C_Prior_TX_V3")
+      .reviewPriorTreatmentForm()
+
+      .sectionBreak("Prostate Diagnosis")
+      .selectFromSidebar('#Prostate_Diagnosis_V4Button')
+      .reviewPageTitle("Prostate_Diagnosis_V4")
+      .reviewProstateDiagnosisForm()
+
+      .sectionBreak("Blood Labs")
+      .selectFromSidebar('#Blood_Labs_V2Button')
+      .reviewPageTitle("Blood_Labs_V2")
+      .reviewBloodLabsForm()
+
+      .sectionBreak("Demographics")
+      .selectFromSidebar('#DemographicsButton')
+      .reviewPageTitle("Demographics")
+      .reviewDemographicsForm()
+
+      .sectionBreak("Weight")
+      .selectFromSidebar('#ECOG_Weight_V3Button')
+      .reviewPageTitle("ECOG_Weight_V3")
+      .reviewWeightForm()
+
+      .sectionBreak("Disease Assessment")
+      .selectFromSidebar('#GU_Disease_Assessment_V3Button')
+      .reviewPageTitle("GU_Disease_Assessment_V3")
+      .reviewDiseaseAssessmentForm()
+
+      .sectionBreak("Prior Ca Treatment")
+      .sectionBreak("Biopsy AE")
+      .selectFromSidebar('#SU2C_Biopsy_AE_V1Button')
+      .reviewPageTitle("SU2C_Biopsy_AE_V1")
+      .reviewBiopsyAeForm()
+
+      .sectionBreak("Prior Ca Treatment")
+      .selectFromSidebar('#SU2C_Pr_Ca_Tx_Sumry_V2Button')
+      .reviewPageTitle("SU2C_Pr_Ca_Tx_Sumry_V2")
+      .reviewPriorCaTreatmentForm()
+
+      .sectionBreak("Specimen")
+      .selectFromSidebar('#SU2C_Specimen_V1Button')
+      .reviewPageTitle("SU2C_Specimen_V1")
+      .reviewSpecimenForm()
+
+      .sectionBreak("Patient Enrollment")
+      .selectFromSidebar('#Patient_Enrollment_formButton')
+      .reviewPageTitle("Patient_Enrollment_form")
+      .reviewPatientEnrollmentForm()
+
+      .sectionBreak("Histology Research")
+      .selectFromSidebar('#Histology_ResearchButton')
+      .reviewPageTitle("Histology_Research")
+      .reviewHistologyResearchForm()
+
+      .sectionBreak("Tissue Specimen")
+      .selectFromSidebar('#Tissue_Specimen_formButton')
+      .reviewPageTitle("Tissue_Specimen_form")
+      .reviewTissueSpecimenForm()
+
+      .sectionBreak("Blood Specimen")
+      .selectFromSidebar('#Blood_Specimen_formButton')
+      .reviewPageTitle("Blood_Specimen_form")
+      .reviewBloodSpecimenForm()
+
+      .sectionBreak("Histology Assessment")
+      .selectFromSidebar('#Histological_Assessment_formButton')
+      .reviewPageTitle("Histological_Assessment_form")
+      .reviewHistologyAssessmentForm()
+
+      .sectionBreak("Laser Capture Microdissection")
+      .selectFromSidebar('#Laser_Capture_MicrodissectionButton')
+      .reviewPageTitle("Laser_Capture_Microdissection")
+      .reviewLaserCaptureForm()
+
+      .sectionBreak("RNA Sequence Completion")
+      .selectFromSidebar('#RNASeq_completion_formButton')
+      .reviewPageTitle("RNASeq_completion_form")
+      .reviewRnaSequenceForm()
+
+      .sectionBreak("Pathology")
+      .selectFromSidebar('#Pathology_formButton')
+      .reviewPageTitle("Pathology_form")
+      .reviewPathologyForm()
+
+      .signOut()
+      .reviewSignInPage()
+      .end();
+  }
+};
